netutils: isolate linux dependent parts
