+++ pkg/util/netutils2/netutils_others.go
@@ -1,0 +1,7 @@
+// +build !linux
+
+package netutils2
+
+func (n *SNetInterface) GetAddresses() [][]string {
+	return nil
+}
